refactor: SSE stream response 패키지 이동
